requirements.txt
aiogram
requests
python-dateutil
pillow
pytesseract
google-genai